+++ src/utils/notifications.ts
@@ -1,24 +1,25 @@
 import { launchCustomDialog } from "./ui/customDialog";
 import { friendlyName } from "../../package.json";
+import { registerServiceWorker } from "./serviceWorkers";
 
 export const checkNotificationPermission = async () => {
   if (!Notification) {
     launchCustomDialog(
       "Notifications",
-      "Your browser does not support desktop notifications.",
+      "Your browser does not support desktop notifications."
     );
   } else {
     switch (Notification.permission) {
       case "granted":
         showNotification(
           friendlyName,
-          "Desktop notifications are already enabled",
+          "Desktop notifications are already enabled"
         );
         break;
       case "denied":
         launchCustomDialog(
           "Notifications",
-          "Permissions have been denied. Please change through your browser settings for this page.",
+          "Permissions have been denied. Please change through your browser settings for this page."
         );
         break;
       case "default":
@@ -34,9 +35,10 @@ const showNotification = (title: string, content: string) =>
 
 const handlePermissionJustGranted = () => {
   if (Notification.permission === "granted") {
+    registerServiceWorker("notification");
     showNotification(
       friendlyName,
-      "Desktop notifications have just been enabled.",
+      "Desktop notifications have just been enabled."
     );
   }
 };