+++ src/webComponents/links/TextLink.ts
@@ -9,7 +9,6 @@ class TextLink extends HTMLElement {
   connectedCallback() {
     const a = this.shadowRoot.querySelector("a");
     a.href = this.dataset.url;
-    a.title = this.dataset.tooltip;
     if (this.dataset.external === "true") {
       a.target = "__blank";
     }