+++ src/serviceWorkers/notificationServiceWorker.ts
@@ -9,7 +9,7 @@ const apiBaseUrl =
 
 const savePushSubscription = async (subscription: PushSubscription) => {
   const res = await fetch(
-    `${apiBaseUrl}/api/subscription?appId=65033d73-3cdc-416d-9689-1b0b3e73cbad`,
+    `${apiBaseUrl}/api/subscription?appId=fe24b4fa-004b-44fb-ac30-9fdb487246c0`,
     {
       method: "POST",
       headers: {
@@ -17,7 +17,7 @@ const savePushSubscription = async (subscription: PushSubscription) => {
         "Content-Type": "application/json",
       },
       body: JSON.stringify(subscription),
-    },
+    }
   );
   return await res.json();
 };
@@ -26,7 +26,16 @@ notificationSW.addEventListener("activate", async () => {
   const subscription = await notificationSW.registration.pushManager.subscribe({
     userVisibleOnly: true,
     applicationServerKey:
-      "BJIUt7Pt_XgjUCatT8Vuu7Tyj0kVaT0t4OaUuWzdLtoINF95-LiI7LR0RzpIBRhmRTp0W9g2dI_OlWWy_OyMMgo",
+      "BEfuj-su_7dqT40eFWTa4wh8FZDJ5oPUiu8AqxFQ260hZotE3i0ZH5B8Esc2J126zJgxLSEKSBRsrtFbKPXRo4Y",
   });
   await savePushSubscription(subscription);
+});
+
+notificationSW.addEventListener("push", async (event) => {
+  const data = event.data?.json();
+
+  await notificationSW.registration.showNotification(data.title, {
+    body: data.body,
+    icon: "/favicon.svg",
+  });
 });