+++ src/utils/ui/ui.ts
@@ -1,8 +1,10 @@
 import type PrettyButton from "../../webComponents/form/PrettyButton";
+import SocialLink from "../../webComponents/links/SocialLink";
+import { checkNotificationPermission } from "../notifications";
 
 export const disableContextMenu = () =>
   document.body.addEventListener("contextmenu", (event) =>
-    event.preventDefault(),
+    event.preventDefault()
   );
 
 export const preventNavigation = () => {
@@ -18,7 +20,7 @@ export const deleteChildren = (parent: HTMLElement) =>
 
 export const setupFullscreenBtn = async () => {
   const fullscreenBtn = document.getElementById(
-    "fullscreenBtn",
+    "fullscreenBtn"
   ) as PrettyButton;
   fullscreenBtn.callback = async () => {
     if (document.fullscreenElement) {
@@ -27,4 +29,11 @@ export const setupFullscreenBtn = async () => {
       await document.body.requestFullscreen();
     }
   };
+};
+
+export const setupNotificationsIcon = () => {
+  const notificationsIcon = document.getElementById(
+    "notificationsIcon"
+  ) as SocialLink;
+  notificationsIcon.action = async () => await checkNotificationPermission();
 };