+++ src/webComponents/links/SocialLink.ts
@@ -10,6 +10,14 @@ class SocialLink extends HTMLElement {
     const a = this.shadowRoot.querySelector("a");
     a.href = this.dataset.url;
     a.title = this.dataset.tooltip ?? "";
+    if (this.dataset.external === "true") {
+      a.target = "_blank";
+      a.rel = "noopener noreferrer";
+    }
+  }
+
+  set action(value: () => void) {
+    this.addEventListener("click", value);
   }
 }