+++ src/main.ts
@@ -4,7 +4,6 @@ import {
   disableContextMenu,
   preventNavigation,
   setupFullscreenBtn,
-  setupNotificationsIcon,
 } from "./utils/ui/ui";
 import { setupKeyboardControls } from "./utils/controls";
 import { registerServiceWorker } from "./utils/serviceWorkers";
@@ -61,6 +60,6 @@ document.addEventListener("DOMContentLoaded", async () => {
 
   setupScoresDialog();
   setupFullscreenBtn();
-  setupNotificationsIcon();
+  //setupNotificationsIcon();
   setupMusic();
 });