+++ changelog.md
@@ -3,7 +3,7 @@
 - Internal:
   - Upgrade to Vite 7.3.1 and change it to a developer dependency
   - Update to Typescript 5.9.3
-  - Refactor `<text-link>` Web Component to read `data-external` attribute
+  - Refactor `<text-link>` Web Component to read `data-external` attribute and improve `data-tooltip`
   - Version printing moved to build-step
   - Remove redundant `ts-loader` package
 - Settings dialog: