+++ changelog.md
@@ -8,6 +8,9 @@
   - Remove redundant `ts-loader` package
 - Settings dialog:
   - Remove About tab
+  - Add Permissions tab
+- Mobile:
+  - Screen for portrait error now occupies full screen height
 
 # 0.8.4 Beta - 06/01/2026 17:45 UTC+1