+++ changelog.md
@@ -11,6 +11,8 @@
   - Add Permissions tab
 - Mobile:
   - Screen for portrait error now occupies full screen height
+- Fullscreen:
+  - Clicking button once already in fullscreen now exits fullscreen mode
 
 # 0.8.4 Beta - 06/01/2026 17:45 UTC+1